day 6
I)i finished input logics, in the right block generates cv,2nd and 3rd pages can create more... and it also generates cv in the right box.
II)started validations, created first 5 validation and each works perfectly..
III) i have one problem and dont know how to get it, the 'logo-12' little icon in right block must be bottom of the page and i made it with position absolute, but if cv is too high , this icon looses the position and i have no idea what is the one position that this little icon can get and be in the bottom left corner everytime... i hope i'll cope with everything.